fix CardList
